+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/DocumentParser_allGraphQLCases_Server_Test.java
@@ -76,7 +76,7 @@ class DocumentParser_allGraphQLCases_Server_Test {
 
 		// Verification
 		assertEquals(44, i, "Nb java files are generated");
-		assertEquals(8, generateCodeDocumentParser.getDirectives().size(), "Nb directives");
+		assertEquals(9, generateCodeDocumentParser.getDirectives().size(), "Nb directives");
 		assertEquals(27, generateCodeDocumentParser.getObjectTypes().size(), "Nb objects");
 		assertEquals(5, generateCodeDocumentParser.getCustomScalars().size(), "Nb custom scalars");
 		assertEquals(14, generateCodeDocumentParser.getInterfaceTypes().size(), "Nb interfaces");
@@ -94,13 +94,13 @@ class DocumentParser_allGraphQLCases_Server_Test {
 		assertTrue(dataFetcher.isCompletableFuture(), "oneWithIdSubType: the dataLoader one");
 		//
 		dataFetcher = findDataFetcher("DataFetchersDelegateAllFieldCases", "listWithIdSubTypes", 1);
-		assertFalse(dataFetcher.isCompletableFuture(), "listWithIdSubTypes (only standard dataFethcher here)");
+		assertTrue(dataFetcher.isCompletableFuture(), "listWithIdSubTypes is annotated by @GenerateDataLoaderForLists");
 		//
 		dataFetcher = findDataFetcher("DataFetchersDelegateAllFieldCases", "oneWithoutIdSubType", 1);
-		assertFalse(dataFetcher.isCompletableFuture(), "oneWithoutIdSubType (only standard dataFethcher here)");
+		assertFalse(dataFetcher.isCompletableFuture(), "oneWithoutIdSubType (only standard dataFetcher here)");
 		//
 		dataFetcher = findDataFetcher("DataFetchersDelegateAllFieldCases", "listWithoutIdSubTypes", 1);
-		assertFalse(dataFetcher.isCompletableFuture(), "listWithoutIdSubTypes (only standard dataFethcher here)");
+		assertFalse(dataFetcher.isCompletableFuture(), "listWithoutIdSubTypes (only standard dataFetcher here)");
 
 		//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 		// Checks if input types for the AllFieldCases object are correctly read
@@ -237,6 +237,7 @@ class DocumentParser_allGraphQLCases_Server_Test {
 		assertEquals("deprecated", generateCodeDocumentParser.getDirectives().get(i++).getName());
 		assertEquals("IDScalarDirective", generateCodeDocumentParser.getDirectives().get(i++).getName());
 		assertEquals("RelayConnection", generateCodeDocumentParser.getDirectives().get(i++).getName());
+		assertEquals("GenerateDataLoaderForLists", generateCodeDocumentParser.getDirectives().get(i++).getName());
 		assertEquals("testDirective", generateCodeDocumentParser.getDirectives().get(i++).getName());
 		assertEquals("anotherTestDirective", generateCodeDocumentParser.getDirectives().get(i++).getName());