+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/impl/AbstractIT.java
@@ -269,7 +269,6 @@ abstract class AbstractIT {
 		assertEquals(nbItemsWithoutId, allFieldCases.getListWithoutIdSubTypes().size());
 		assertTrue(
 				allFieldCases.getListWithoutIdSubTypes().get(0).getName().endsWith(textToAppendToTheFornameWithoutId));
-
 	}
 
 	@Execution(ExecutionMode.CONCURRENT)